% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\docType{data}
\name{clean_df}
\alias{clean_df}
\title{clean_df obtained from the RFID / PIT tag readers}
\format{
A data frame with xxx rows and yy variables:
\describe{
  \item{id}{RFID /PIT tag id of individuals}
  \item{time}{time of read}
  \item{antenna}{antenna number}
  ...
}
}
\usage{
clean_df
}
\description{
A dataset containing the putput with required formatted columns
"id", "time", "date", "antenna".
}
\keyword{datasets}
