% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bold_targets.R
\name{filter_bold_targets}
\alias{filter_bold_targets}
\title{Filter all the boldness targets}
\usage{
filter_bold_targets(mydataset)
}
\arguments{
\item{mydataset}{Dataframe with the PIT tag data. It should already be filtered according to the time of exploration/boldness experiments}
}
\value{
The filtered dataset
}
\description{
In these pond experiments, boldness targets (4 targets at the top) were kept the same throughout the experiments
The antenna numbers (or `Unit number` in the output file) that corresponds to boldness targets are: 11,12,13,14,31,32,33,35
This function filters only the boldness targets from the given dataset
}
