% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_exp_targets.R
\name{filter_exp_targets}
\alias{filter_exp_targets}
\title{Filter all the exploration targets}
\usage{
filter_exp_targets(mydataset)
}
\arguments{
\item{mydataset}{Dataframe with the PIT tag data. It should already be filtered according to the time of exploration experiments}
}
\value{
The filtered dataset
}
\description{
In these pond experiments, exploration targets (5 targets at the bottom) were kept the same throughout the experiments
The antenna numbers (or `Unit number` in the output file) that corresponds to exploration targets are: 21,22,23,24,25,41,42,43,44,45
This function filters only the exploration targets from the given dataset
}
