% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_crosses.R
\name{get_all_crosses}
\alias{get_all_crosses}
\title{Get all crosses}
\usage{
get_all_crosses(pit_dataset, test_details)
}
\arguments{
\item{mydataset}{Given dataset with start and endtimes for filtering, filter tests}
}
\value{
The filtered dataset
}
\description{
pit_dataset is the cleaned up dataset
}
