% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_data.R
\name{get_clean_data}
\alias{get_clean_data}
\title{clean_up}
\usage{
get_clean_data(merged_dataset)
}
\arguments{
\item{merged_dataset}{A dataframe after merging all the excel files from PIT tag data}
}
\value{
a cleaned dataframe ready for getting useful cols
}
\description{
Sometimes a little bit of previous day's data is included and
hence we have to make sure that we
remove any duplicates of Identifier
(the first column, which is the serial number of recording
that the data logger keeps tracks of )
}
