% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crosses.R
\name{get_crosses}
\alias{get_crosses}
\title{Get all crosses}
\usage{
get_crosses(input_dataset)
}
\arguments{
\item{mydataset}{Given dataset with start and endtimes for filtering, filter tests}
}
\value{
The filtered dataset
}
\description{
Get all crosses
}
