% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hourly_crosses.R
\name{get_hourly_crosses}
\alias{get_hourly_crosses}
\title{Get hourly crosses}
\usage{
get_hourly_crosses(
  pit_data,
  test_details,
  method = c("exp", "pond"),
  total_pond_hours = 16,
  total_exp_hours = 5
)
}
\arguments{
\item{pit_data}{A dataframe: data that you get from the PIT antenna}

\item{test_details}{Dataframe containing the details of the test including the columns "test_ID","Start_time_exp","End_time_exp", "Start_time_pond","End_time_pond"}
}
\value{
a datafrme with crosses between antenna per hour
}
\description{
pit_data is the data you get directly from the PIT tag readers. It can be for one or multiple tests.
  This function is particularly for handling exploration_migration tests in the ponds
}
