% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latency_antenna.R
\name{get_latency_antenna}
\alias{get_latency_antenna}
\title{Get latency : time until an antenna was 'crossed' by an individual}
\usage{
get_latency_antenna(pit_data, test_details)
}
\arguments{
\item{pit_data}{A dataframe: data that you get from the PIT antenna}

\item{test_details}{Dataframe containing the details of the test including the columns "test_ID","Start_time_exp","End_time_exp", "Start_time_pond","End_time_pond"}
}
\value{
a dataframe with antenna number and the time of first detection at the antenna
}
\description{
pit_data is the data you get directly from the PIT tag readers. It can be for one or multiple tests.
  This function is particularly for handling exploration_migration tests in the ponds
}
