% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pond_duration.R
\name{get_pond_duration}
\alias{get_pond_duration}
\title{Get time spent in ponds}
\usage{
get_pond_duration(pit_data, test_details)
}
\arguments{
\item{pit_data}{A dataframe: data that you get from the PIT antenna}

\item{test_details}{Dataset with details of the test. It should contain the columns "test_id", "start_time", "end_time"}
}
\value{
The filtered dataset
}
\description{
pit_data is the cleaned up dataset
}
