% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_useful_cols.R
\name{get_useful_cols}
\alias{get_useful_cols}
\title{Get only the useful columns of the dataset}
\usage{
get_useful_cols(dataset)
}
\arguments{
\item{dataset}{A dataframe: data that you get from the PIT antenna}
}
\value{
The same dataframe but without the unnecessary columns
}
\description{
This function trims the dataset from the original raw data obtained from PIT antenna
}
\details{
The function can dataframe of any kind
}
