% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_ref_df.R
\docType{data}
\name{id_ref_df}
\alias{id_ref_df}
\title{id_ref frame obtained from the RFID / PIT tag readers}
\format{
A data frame with xxx rows and yy variables:
\describe{
  \item{id}{RFID /PIT tag id of individuals}
  \item{Date}{date of read}
  \item{SL}{time of read}
  \item{Weight}{antenna number}
  \item{Plates}{Plates}
  \item{Tagger}{Jakob}
  \item{Remarks}{Remarks while tagging}
  \item{Pond}{pond_id}
  \item{Treatment_seq}{treatment sequence}
  \item{Replicate}{replicate number}
}
}
\usage{
id_ref_df
}
\description{
A dataset containing the putput with required formatted columns
"id", "time", "date", "antenna".
}
\keyword{datasets}
