% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimerge.R
\name{multimerge}
\alias{multimerge}
\title{Merge all csv files in a folder}
\usage{
multimerge(mypath, extension)
}
\arguments{
\item{mypath}{Path for the folder containing the datafiles.}

\item{extension}{File extension.}
}
\value{
The combined dataset
}
\description{
This function merges all the csv files that we get from the TROVAN PIT tag system.
It requires that all the datafiles from the same experiment be placed in a single folder
}
\details{

}
\note{
Path should be with 'my/path/' or "my\\path\\". Here only .csv works for now.
}
\examples{
mydata<-multimerge("C:\\\\Users\\\\User\\\\Documents\\\\GitHub\\\\Personality-in-semi-natural-ponds\\\\data\\\\raw_data_mixed","*.CSV")

}
