% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_clean_data.R
\name{pr_clean_data}
\alias{pr_clean_data}
\title{Clean the raw data from the readers}
\usage{
pr_clean_data(raw_df, id_ref_df)
}
\arguments{
\item{raw_df}{Raw data frame from the PIT / RFID system,
already formatted and containing the columns
"time", "date", "antenna", "id"  (already formatted). This data frame corresponds to
the reads for \emph{one} experimental block.}

\item{id_ref_df}{Reference data frame for an experimental block
containing at least one column "id", with  all
individuals and corresponding variables (already formatted)}
}
\value{
A data frame (\code{clean_df}), with three columns (\emph{i.e.} id, antenna, time),
and the number of rows corresponding to the total number
of reads except duplicates within a second.
}
\description{
\itemize{The function does the following
\item{}{Keeps columns of interest (i.e. id, antenna, time)}
\item{}{Posixt format for time}
\item{}{Keep one read per individual at a certain antenna per second}
\item{}{Remove ghost reads}
\item{}{Orders the data frame based on time}

}
}
