% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_dist.R
\name{pr_dist}
\alias{pr_dist}
\title{Distance travelled by an individual}
\usage{
pr_dist(block_df, block_ref_df, ant_coordinates)
}
\arguments{
\item{block_df}{A data frame containing the reads from an experimental block}

\item{block_ref_df}{A data frame containing a reference list with all individuals present in the experimental block.}

\item{ant_coordinates}{A data frame containing  four columns: "antenna","x","y","z" (where "x", "y", and "z" represent the coordinates in three domensions). If there is no z axis, enter all zeros but do not omit the column.}
}
\value{
A data frame containing two columns: "id", and "distance", indicating the travelled distance for every individual.
}
\description{
This function computes, for every individual, the euclidean distance travelled between all antennas it has been read at.
}
