% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_sociality.R
\name{pr_sociality}
\alias{pr_sociality}
\title{Computes the proportion of time spent with at least one conspecific}
\usage{
pr_sociality(block_df, block_ref_df, cutoff = 10)
}
\arguments{
\item{block_df}{A data frame containing the reads from an experimental block}

\item{block_ref_df}{A data frame containing a reference list with all individuals present in the experimental block.}

\item{cutoff}{A value corresponding to the minimum number of \emph{total reads} (both accompanied and non-accompanied) from which a
sociability index (\emph{i.e.} proportion of accompanied reads) should be computed. For example, if an individual has been
read only once, by chance, in presence of a conspecific, a score of 100\% for this individual is likely non-representative of the
real proportion of time it spends in presence of conspecifics By default, the value if set at 10 total reads. For no cutoff, one can
use a value of 0.}
}
\value{
A data table containing (i) the number of accompanied reads (\emph{i.e.} reads for which at least
another individual was present at the same antenna and at the same second); (ii) total number of reads;
(iii) proportion of accompanied reads (\code{proportion} = \code{acc_reads} / \code{tot_reads}).
}
\description{
The function computes, for every individual, the proportion of time it has been read in presence of at least
one conspecific (read at the same second at the same antenna).
}
\examples{

data(block_df)
data(block_ref_df)


pr_sociality(block_df,
            block_ref_df)

pr_sociality(block_df,
            block_ref_df,
            cutoff = 0)


}
