% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_df.R
\docType{data}
\name{raw_df}
\alias{raw_df}
\title{Raw data frame obtained from the RFID / PIT tag readers}
\format{
A data frame with 808 rows and yy variables:
\describe{
  \item{id}{RFID /PIT tag id of individuals}
  \item{date}{date of read}
  \item{time}{time of read}
  \item{antenna}{antenna number}
  \item{Antenna.number}{time of read}
  \item{Event}{time of read}
  \item{GPS.coordinates}{time of read}
  \item{Identifier}{time of read}
  \item{Input.status}{time of read}
  \item{Output.status}{time of read}
  \item{Transponder.type}{time of read}
  \item{Weight}{time of read}
  ...
}
}
\usage{
raw_df
}
\description{
A dataset containing the putput with required formatted columns
"id", "time", "date", "antenna".
}
\keyword{datasets}
