% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pond_duration.R
\name{get_pond_duration}
\alias{get_pond_duration}
\title{Get time spent in ponds}
\usage{
get_pond_duration(pit_pond, test_details)
}
\arguments{
\item{pit_pond}{Given dataset with start and endtimes for filtering, filter tests}

\item{test_details}{Dataset with details of the test. It should contain the columns "test_id", "start_time", "end_time"}
}
\value{
The filtered dataset
}
\description{
pit_dataset is the cleaned up dataset
}
