% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_duration.R
\name{get_read_duration}
\alias{get_read_duration}
\title{Get read duration
Duration of a read in antenna}
\usage{
get_read_duration(pit_data, test_details)
}
\arguments{
\item{pit_data}{A dataframe: data that you get from the PIT antenna}

\item{test_details}{Dataframe containing the details of the test including the columns "test_ID","Start_time_exp","End_time_exp", "Start_time_pond","End_time_pond"}
}
\value{
a datafrme with 5 columns containing tag_ID, test_id, cross, antenna, duration at the antenna
}
\description{
pit_data is the data you get directly from the PIT tag readers. It can be for one or multiple tests.
  This function is for getting time an individual spent at any antenna - hiding, taking shelter, crossing antenna etc
}
