% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_dist.R
\name{pr_dist}
\alias{pr_dist}
\title{Distance travelled by an individual}
\usage{
pr_dist(block_df, block_ref_df, ant_coordinates)
}
\arguments{
\item{block_df}{A data frame containing the reads from an experimental block}

\item{block_ref_df}{A data frame containing a reference list with all individuals present in the experimental block.}

\item{ant_coordinates}{A data frame containing  "antenna" and coordinates, "x","y","z" in other columns. If there is no z axis, enter all zeros but do not omit the column}
}
\value{
A data frame containing the "id" and corresponding "distance" travelled by fish
}
\description{
Distance travelled: eucleadian distance
travelled by fish: block_df, block_ref_id_df,
block_ref_antenna_df, output gives indiv, distance
travelled, 0 for indiv that got read, didnt move and
 NA for indiv that did not get read
}
\examples{

}
