% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_summary.R
\name{pr_summary}
\alias{pr_summary}
\title{pr_summary Function}
\usage{
pr_summary(block_df, block_ref_df)
}
\arguments{
\item{block_ref_df}{}
}
\value{

}
\description{
This function takes a data frame containing the reads of an experimental block and produces (i)
a summary table containing the reads of each individual per antenna, and the total reads;
(ii) a heatmap corresponding to a visual representation of the table.
}
\examples{


}
