% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_summary.R
\name{pr_summary}
\alias{pr_summary}
\title{pr_summary Function
#'}
\usage{
pr_summary(block_df, block_ref_df)
}
\arguments{
\item{block_df}{Raw data frame from the PIT / RFID system,
already formatted and containing these columns,
"time", "date", "antenna", "id"  (already formatted)}

\item{block_ref_df}{Reference data frame from the whole experiment
containing at least one column "id", with  all
individuals and corresponding variables (already formatted)}
}
\value{
A data frame (clean_df), with three columns (i.e. id, antenna, time),
and the number of rows corresponding to the total number
of reads except duplicates within a second.
}
\description{
\itemize{The function does the following:
\item Remove ghost reads
\item Keep one read per individual at a certain antenna per second
\item Posixt format for time
\item Keeps columns of interest (i.e. id, antenna, time)}
}
